+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   `getRanges()`, `getNetworks()`, `isAllowed()` methods (@vjik)
 - Enh #746: Use `NEGATION_CHARACTER` constant from `network-utilities` package in `IpHandler` instead of declaring its own 
   constant (@arogachev)
+- Chg #747: Merge rules from PHP attributes with rules provided via `getRules()` method (@vjik)
 
 ## 2.0.0 August 02, 2024